Split phase4 module into cloud_io / lidar / fusion topic modules
Module names should describe what they do, not what planning phase they
came from. Replaced phase4.py with three topic-named modules:

* digitalrivers.cloud_io — tile_windows, write_cog, plus the
  dask_backend and cloud_storage umbrella stubs for the deferred
  full-Dask-graph and Zarr / S3 / GCS factories.
* digitalrivers.lidar — grid_lidar_points and the pdal_lidar_pipeline
  umbrella stub for the deferred full LAS read + classify + filter
  pipeline.
* digitalrivers.fusion — topobathy_fusion (four blend modes:
  max / min / topo_above / bathy_below).

Dropped the four pointer-style umbrella stubs that did nothing but
raise NotImplementedError to redirect callers at an existing working
public API:

* native_cotat_upscale → FlowDirection.upscale(method='cotat')
* native_ihu_upscale → FlowDirection.upscale_ihu
* anudem_solver → DEM.anudem_interpolate
* mesh_quality_optimise → Mesh.laplacian_smooth

The working public methods already document themselves; the umbrella
stubs were pure cruft and made the module surface confusing.

Renamed cloud_io umbrella (originally cloud_io function in phase4) to
cloud_storage to avoid a function/module name collision.

Renamed tests:
* test_grid_lidar.py → test_lidar.py (now also covers pdal_lidar_pipeline)
* test_phase4_stubs.py → split into test_cloud_io.py + test_fusion.py
* test_phase4_coverage_gaps.py → test_phase4_review_fixups.py
  (test files keep "phase-4" only as historical review-pass context).

522 unit tests + 4 doctests pass.
